+++ tests/cypress/latest/e2e/unit_tests/capd_rke2_cluster.spec.ts
@@ -18,14 +18,16 @@ import { skipClusterDeletion } from '~/support/utils';
 
 Cypress.config();
 describe('Import CAPD RKE2', { tags: '@short' }, () => {
+  var clusterName: string, clusterPrefix: string
   const timeout = 300000
-  const classesRepo = 'classes'
-  const clustersRepo = 'clusters'
-  const clusterName = 'cluster1'
-  const className = 'quick-start'
+  const classesPath = 'classes'
+  const clustersPath = 'clusters'
+  const className = 'capd-rke2-class'
+  const clusterNamePrefix = 'capd-rke2-cluster' // as per fleet values
+  const classClusterNamePrefix = className + '-cluster' // as per fleet values
   const repoUrl = 'https://github.com/rancher/rancher-turtles-e2e.git'
   const basePath = '/tests/assets/rancher-turtles-fleet-example/capd/rke2/'
-  const pathNames = ['clusters', 'clusterclass']
+  const pathNames = [clustersPath, 'clusterclass']
   const branch = 'main'
   const questions = [{ menuEntry: 'Rancher Turtles Features Settings', inputBoxTitle: 'Kubectl Image', inputBoxValue: 'registry.k8s.io/kubernetes/kubectl:v1.31.0' }];
 
@@ -35,31 +37,45 @@ describe('Import CAPD RKE2', { tags: '@short' }, () => {
   });
 
   pathNames.forEach((path) => {
+    const clustersRepoName = clusterNamePrefix + path
+    const classesRepoName = classClusterNamePrefix + path
 
     it('Setup the namespace for importing', () => {
-      if (path.includes('clusters')) {
+      if (path.includes(clustersPath)) {
         cy.namespaceAutoImport('Enable');
       } else {
         cy.namespaceAutoImport('Disable');
       }
     })
 
-    it('Add CAPD cluster fleet repo - ' + path, () => {
+    it('Add CAPD cluster fleet repo(s) - ' + path + ' and get cluster name', () => {
       cypressLib.checkNavIcon('cluster-management').should('exist');
       var fullPath = basePath + path
 
       if (path.includes('clusterclass')) {
-        // Add classes fleet repo to fleel-local workspace
-        fullPath = fullPath.concat('/', classesRepo)
-        cy.addFleetGitRepo(classesRepo, repoUrl, branch, fullPath);
-        fullPath = fullPath.replace(classesRepo, clustersRepo);
+        // Add classes fleet repo to fleet-local workspace
+        fullPath = fullPath.concat('/', classesPath)
+        cy.addFleetGitRepo(classesRepoName, repoUrl, branch, fullPath);
+        fullPath = fullPath.replace(classesPath, clustersPath);
         cypressLib.burgerMenuToggle();
       }
+      cy.addFleetGitRepo(clustersRepoName, repoUrl, branch, fullPath);
 
-      cy.addFleetGitRepo(clustersRepo, repoUrl, branch, fullPath);
+      if (path.includes(clustersPath)) {
+        clusterPrefix = clusterNamePrefix
+      } else {
+        clusterPrefix = classClusterNamePrefix
+      }
+      // Check CAPI cluster using its name prefix
+      cy.checkCAPICluster(clusterPrefix);
+      // Get the cluster name by its prefix and use it across the test
+      cy.getBySel('sortable-cell-0-1').then(($cell) => {
+        clusterName = $cell.text();
+        cy.log('CAPI Cluster Name:', clusterName);
+      });
     })
 
-    if (path == 'rke2_namespace_autoimport') { var qase_id = 29 } else { qase_id = 30 }
+    if (path == clustersPath) { var qase_id = 29 } else { qase_id = 30 }
     qase(qase_id,
       it('Auto import child CAPD cluster', () => {
         // Check child cluster is created and auto-imported
@@ -74,7 +90,7 @@ describe('Import CAPD RKE2', { tags: '@short' }, () => {
       })
     );
 
-    if (path.includes('clusters')) {
+    if (path.includes(clustersPath)) {
       qase(7,
         it('Install App on imported cluster', { retries: 1 }, () => {
           // Click on imported CAPD cluster
@@ -90,6 +106,7 @@ describe('Import CAPD RKE2', { tags: '@short' }, () => {
           // Access CAPI cluster
           cy.checkCAPIMenu();
           cy.contains('Machine Deployments').click();
+          cy.typeInFilter(clusterName);
           cy.getBySel('sortable-table-0-action-button').click();
           cy.contains('Edit YAML')
             .click();
@@ -103,6 +120,7 @@ describe('Import CAPD RKE2', { tags: '@short' }, () => {
 
           // Check CAPI cluster status
           cy.contains('Machine Deployments').click();
+          cy.typeInFilter(clusterName);
           cy.get('.content > .count', { timeout: timeout }).should('have.text', '2');
           cy.checkCAPIClusterProvisioned(clusterName);
         })
@@ -120,47 +138,49 @@ describe('Import CAPD RKE2', { tags: '@short' }, () => {
       );
     }
 
-    qase(9,
-      it('Remove imported CAPD cluster from Rancher Manager', { retries: 1 }, () => {
-        // Check cluster is not deleted after removal
-        cy.deleteCluster(clusterName);
-        cy.goToHome();
-        // kubectl get clusters.cluster.x-k8s.io
-        // This is checked by ensuring the cluster is not available in navigation menu
-        cy.contains(clusterName).should('not.exist');
-        cy.checkCAPIClusterProvisioned(clusterName);
-      })
-    );
-
-    qase(10,
-      it('Delete the CAPD cluster fleet repo(s) - ' + path, () => {
-        if (path.includes('clusterclass')) {
-          // Remove the classes fleet repo
-          cy.removeFleetGitRepo(classesRepo, true);
-          // Remove the clusters fleet repo
-          cypressLib.burgerMenuToggle();
-          cy.removeFleetGitRepo(clustersRepo);
-
-          // Wait until the following returns no clusters found
-          cy.checkCAPIClusterDeleted(clusterName, timeout);
-          // Remove the clusterclass
-          cy.removeCAPIResource('Cluster Classes', className);
-        } else {
-          // Remove the clusters fleet repo
-          cy.removeFleetGitRepo(clustersRepo);
-
-          // Wait until the following returns no clusters found
-          // This is checked by ensuring the cluster is not available in CAPI menu
-          cy.checkCAPIClusterDeleted(clusterName, timeout);
-        }
-
-        // Ensure the cluster is not available in navigation menu
-        cy.getBySel('side-menu').then(($menu) => {
-          if ($menu.text().includes(clusterName)) {
-            cy.deleteCluster(clusterName);
-          }
+    if (skipClusterDeletion) {
+      qase(9,
+        it('Remove imported CAPD cluster from Rancher Manager', { retries: 1 }, () => {
+          // Check cluster is not deleted after removal
+          cy.deleteCluster(clusterName);
+          cy.goToHome();
+          // kubectl get clusters.cluster.x-k8s.io
+          // This is checked by ensuring the cluster is not available in navigation menu
+          cy.contains(clusterName).should('not.exist');
+          cy.checkCAPIClusterProvisioned(clusterName);
         })
-      })
-    );
-  })
+      );
+
+      qase(10,
+        it('Delete the CAPD cluster fleet repo(s) - ' + path, () => {
+          if (path.includes('clusterclass')) {
+            // Remove the classes fleet repo
+            cy.removeFleetGitRepo(classesRepoName, true);
+            // Remove the clusters fleet repo
+            cypressLib.burgerMenuToggle();
+            cy.removeFleetGitRepo(clustersRepoName);
+
+            // Wait until the following returns no clusters found
+            cy.checkCAPIClusterDeleted(clusterName, timeout);
+            // Remove the clusterclass
+            cy.removeCAPIResource('Cluster Classes', className);
+          } else {
+            // Remove the clusters fleet repo
+            cy.removeFleetGitRepo(clustersRepoName);
+
+            // Wait until the following returns no clusters found
+            // This is checked by ensuring the cluster is not available in CAPI menu
+            cy.checkCAPIClusterDeleted(clusterName, timeout);
+          }
+
+          // Ensure the cluster is not available in navigation menu
+          cy.getBySel('side-menu').then(($menu) => {
+            if ($menu.text().includes(clusterName)) {
+              cy.deleteCluster(clusterName);
+            }
+          })
+        })
+      );
+    }
+  })  
 });
